picotemp/requirements.txt
pika==1.3.2
pyserial==3.5
